+++ packages/stream_chat/lib/src/core/platform_detector/platform_detector_stub.dart
@@ -1,7 +1,0 @@
-import 'platform_detector.dart';
-
-/// Stub implementation
-PlatformType get currentPlatform => throw UnimplementedError();
-
-/// Stub implementation
-bool get isFlutterTestEnvironment => throw UnimplementedError();